requirements.txt
yt-dlp
openai-whisper
anthropic
openai
python-dotenv
requests
tqdm
pyannote.audio
torch
pydub
numpy
google-genai
